drm/msm/sde: Introduce KCAL color control
Use PCC and HSIC interfaces to manually calibrate the colors of the
DSI panel via sysfs.  This is a simple driver which uses predefined
API to do this. As it is a purely platform driver, sysfs is located
at /sys/devices/platform/kcal_ctrl.0

Supported PCC and HSIC features:
- RGB control,
- Color distortion control,
- Saturation intensity control,
- Contrast intensity control,
- Color brightness control,
- Grayscale mode,
- Astronomy mode;

We already have atomic ops to alter DSI colors, but users prefer
this interface over that.

 Conflicts:
	drivers/gpu/drm/msm/sde/sde_hw_reg_dma_v1_color_proc.c

Signed-off-by: Alesaiko <solcmdr@gmail.com>
Signed-off-by: Hemant Sachdeva <hemant.evolver@gmail.com>